refactor send endpoint
